+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/DefaultHelpCommand.java
@@ -8,21 +8,18 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-// TODO: 03.08.2018 Docs 
 /**
- * 
+ * The default implementation for AbstractHelpCommand.
+ * If you want to use this, add a new instance of this class as a command in your CommandSettings with the put-method.
+ * This class is final. To create your own help command implementation, please refer to AbstractHelpCommand.
+ * @author JohnnyJayJay
+ * @version
+ * @see AbstractHelpCommand
  */
 public final class DefaultHelpCommand extends AbstractHelpCommand {
 
-    public DefaultHelpCommand(CommandSettings settings) {
-        super(settings);
-    }
-
     /**
-     * 
-     * @param event
-     * @param prefix
-     * @param commands
+     * Lists all commands along with the information that more help can be received by adding the optional label parameter.
      */
     @Override
     public void provideGeneralHelp(CommandEvent event, String prefix, Map<String, ICommand> commands) {
@@ -30,24 +27,21 @@ public final class DefaultHelpCommand extends AbstractHelpCommand {
         if (!selfMember.hasPermission(event.getChannel(), Permission.MESSAGE_WRITE))
             return;
 
-        EmbedBuilder embed = getBuilder(selfMember);
-        String helpLabels = "[" + String.join("|", settings.getLabels(this)) + "]";
+        EmbedBuilder embed = getBuilder(selfMember, event.getCommandSettings());
+        // There may be problems here, since it is possible to add this command on another CommandSettings-instance than the attribute of this class.
+        String helpLabels = "[" + String.join("|", event.getCommandSettings().getLabels(this)) + "]";
         embed.appendDescription("To learn more about a specific command, just call `").appendDescription(prefix)
                 .appendDescription(helpLabels).appendDescription(" <label>`.\nThe following commands are currently available:\n");
         String commandsList = commands.keySet().stream().map((label) -> prefix + label).collect(Collectors.joining(", "));
         if (commandsList.length() < 1010)
-            embed.addField("Commands", "```\n" + commands + "```", false);
+            embed.addField("Commands", "```\n" + commandsList + "```", false);
         else
             embed.addField("Warning", "Too many commands to show.", false);
         event.getChannel().sendMessage(embed.build()).queue();
     }
 
     /**
-     * 
-     * @param event
-     * @param prefix
-     * @param command
-     * @param labels
+     * Shows the command info based on the method ICommand#info in an embed.
      */
     @Override
     public void provideSpecificHelp(CommandEvent event, String prefix, ICommand command, Set<String> labels) {
@@ -55,25 +49,24 @@ public final class DefaultHelpCommand extends AbstractHelpCommand {
         if (!selfMember.hasPermission(event.getChannel(), Permission.MESSAGE_WRITE))
             return;
 
-        EmbedBuilder embed = getBuilder(selfMember);
+        CommandSettings settings = event.getCommandSettings();
+        EmbedBuilder embed = getBuilder(selfMember, settings);
         embed.appendDescription("**Command Info for:** `").appendDescription("[")
                 .appendDescription(String.join("|", settings.getLabels(command))).appendDescription("]`\n\n")
                 .appendDescription(settings.getCommands().get(event.getArgs()[0]).info(event.getMember()));
         event.getChannel().sendMessage(embed.build()).queue();
     }
 
-    private EmbedBuilder getBuilder(Member selfMember) {
+    private EmbedBuilder getBuilder(Member selfMember, CommandSettings settings) {
         return new EmbedBuilder().setColor(settings.getHelpColor() != null ? settings.getHelpColor() : selfMember.getColor());
     }
 
     /**
-     * 
-     * @param member
-     * @return
+     * Returns the default info for this command.
+     * @return "Shows all available commands or provides help for a specific command. Usage: `help <label>`"
      */
     @Override
     public String info(Member member) {
-        return "Shows all available commands or provides help for a specific command. Usage: `"
-                + this.settings.getPrefix(member.getGuild().getIdLong()) + "help <label>`";
+        return "Shows all available commands or provides help for a specific command. Usage: `help <label>`";
     }
 }